Fix : UserTranslatedMessageCell 파일 reference 문제 해결
Issue #346
