--- iOS/PapagoTalk/PapagoTalk/Chat/View/MessageCell/TranslatedMessageCellByUser.swift
@@ -1,19 +1,0 @@
-//
-//  UserTranslatedMessageCell.swift
-//  PapagoTalk
-//
-//  Created by 송민관 on 2020/12/10.
-//
-
-import UIKit
-
-final class UserTranslatedMessageCell: UICollectionViewCell {
-    
-}
-
-extension UserTranslatedMessageCell: MessageCell {
-    func configureMessageCell(message: Message) {
-        configureMessage(of: messageTextView, with: message.text)
-        configureTime(of: timeLabel, with: message.time, shouldShow: message.shouldTimeShow)
-    }
-}
