+++ iOS/PapagoTalk/PapagoTalk/Chat/View/MessageCell/UserTranslatedMessageCell.swift
@@ -5,4 +5,17 @@
 //  Created by 송민관 on 2020/12/10.
 //
 
-import Foundation
+import UIKit
+
+final class UserTranslatedMessageCell: UICollectionViewCell {
+    
+    @IBOutlet private weak var messageTextView: MessageText!
+    @IBOutlet private weak var timeLabel: UILabel!
+}
+
+extension UserTranslatedMessageCell: MessageCell {
+    func configureMessageCell(message: Message) {
+        configureMessage(of: messageTextView, with: message.text)
+        configureTime(of: timeLabel, with: message.time, shouldShow: message.shouldTimeShow)
+    }
+}
